Remove mesher_base: generate method need finalize keyword for GmshMesher
